Add Asciidoctor to Travis
Set Travis to generate HTML files from AsciiDoc files in docs using
Asciidoctor. Generated files are then pushed to gh-pages branch for
commits to the master branch.

To be able to push files to the repo, Travis needs a deploy key.
There are a few ways of managing deploy keys, as described in
https://developer.github.com/guides/managing-deploy-keys/. This repo
uses HTTPS cloning with OAuth tokens as it is the simplest to set up.
A drawback of this approach is that a token grants access to all of the
user's repos. This means that any member with write access to the repo
is able to change the deploy script to print the token and hence gain
access to the user's repo. Hence, a machine user (with write access to
the repo) should be created instead, and used to generate the personal
access token. This way, only the repo is put at risk in the case of a
leak. While SSH deploy keys are repo-specific, encrypting the key
requires installing Travis Command Line Client which can be
inconvenient.

In deploy_github_pages.sh, note that GitHub prints the repo url when
there is an error for certain commands. This means that the GitHub token
would be printed to Travis logs. To prevent this, we redirect the output
to the null device so that it is not displayed. However, this means that
in case of an error, no output is displayed for debugging. If output is
required, redirection can be temporarily stopped. The GitHub token should
then be changed to a new one.